ИСправление логики класса Player
