+++ Game/headers/Player.h
@@ -14,7 +14,7 @@ public:
 	//Размещение корабля на карте
 	bool setShip(const Cords& c, const int& _dir, const int& _deck)noexcept;
 	//Проверка выстрела
-	std::unique_ptr<Player>&& setShot(std::unique_ptr<Player>&& plr, Cords& cord, const int& pol_count)noexcept;
+	//std::unique_ptr<Player>&& setShot(std::unique_ptr<Player>&& plr, Cords& cord, const int& pol_count)noexcept;
 	//Проверка попадания
 	bool getShot(Cords& crd);
 	//Проверка пересечения кораблей
@@ -30,6 +30,18 @@ public:
 	//Для проверки работы
 	void print()const;
 
+
+
+
+	std::vector<Ship> getNavy() { return navy; }
+
+	void setPoint(Cords& crd, size_t pol, char point) {
+		int count = 0;
+		count = ((crd.second - 1) * static_cast<int>(pol) + crd.first) - 1;
+		map_shot[count] = point;
+	}
+
+	void moveStat(bool m) { move = m; }
 };
